fix(types): resolve all mypy static analysis errors
- Add missing return type annotations across the codebase
- Use async_sessionmaker for correct SQLAlchemy async typing
- Add runtime assertions in ReplayService for type narrowing
- Fix variable shadowing in main.py
- Add explicit casts in middleware and container for strict typing
- Install and configure type stubs

+++ src/rsw/models/degradation/rls.py
@@ -153,7 +153,7 @@ class RLSEstimator:
         """Get root mean squared error of updates."""
         if self.n_updates == 0:
             return 0.0
-        return np.sqrt(self.residual_sum / self.n_updates)
+        return float(np.sqrt(self.residual_sum / self.n_updates))
     
     def reset(self, initial_covariance: float = 1000.0) -> None:
         """Reset the estimator to initial state."""
